refactor(server): define package-owned route wrappers
