+++ packages/doc-server/src/agent.ts
@@ -1,1 +1,8 @@
-export { agentRoutes } from '../../../server/agent-routes.js';
+import type { Router } from 'express';
+import { agentRoutes } from '../../../server/agent-routes.js';
+
+export function createAgentRouter(): Router {
+  return agentRoutes;
+}
+
+export { agentRoutes };
