+++ packages/doc-server/src/bridge.ts
@@ -1,1 +1,8 @@
-export { bridgeRouter, createBridgeMountRouter } from '../../../server/bridge.js';
+import type { Router } from 'express';
+import { bridgeRouter, createBridgeMountRouter } from '../../../server/bridge.js';
+
+export function createBridgeRouter(): Router {
+  return bridgeRouter;
+}
+
+export { bridgeRouter, createBridgeMountRouter };
