+++ packages/doc-server/src/share.ts
@@ -1,1 +1,8 @@
-export { shareWebRoutes } from '../../../server/share-web-routes.js';
+import type { Router } from 'express';
+import { shareWebRoutes } from '../../../server/share-web-routes.js';
+
+export function createShareRouter(): Router {
+  return shareWebRoutes;
+}
+
+export { shareWebRoutes };
